refactor(LottoGameResultGenerator): profits 자료형 변경
int의 크기제한으로 인한 오류가 우려됨. 따라서 profits(수익금)의 자료형을 long으로 변경

+++ src/main/java/lotto/domain/generator/LottoGameResultGenerator.java
@@ -85,12 +85,12 @@ public class LottoGameResultGenerator {
     }
 
     public void calculateEarningRate(int paidMoney) {
-        int profits = calculateProfits();
+        long profits = calculateProfits();
         earningRate = (profits / (double)paidMoney) * 100;
     }
 
-    public int calculateProfits() {
-        int profits = 0;
+    public long calculateProfits() {
+        long profits = 0;
         for (int prizeMoney : winningDetails.keySet()) {
             profits += prizeMoney * winningDetails.get(prizeMoney);
         }
